pyproject.toml
[tool.poetry]
name = "notify-family-budget"
version = "0.1.0"
description = "dev env for notify-family-budget"
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.7"
selenium = "4.0.0"
python-dotenv = "0.14.0"
pandas = "1.1.4"
requests = "2.26.0"
chromedriver-binary = "105.0.*"
ipdb = "^0.13.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
